Some parameters have been changed to their correct type.

+++ code/CoolJetpackWeatherApp/app/src/main/java/dam_A51696/cooljetpackweatherapp/ui/WeatherCard.kt
@@ -12,7 +12,7 @@ import androidx.compose.ui.unit.dp
 fun WeatherCard(
     temperature: Float,
     windSpeed: Float,
-    windDirection: Int,
+    windDirection: Float,
     seaLevelPressure: Float,
     time: String,
     modifier: Modifier = Modifier
