+++ code/CoolJetpackWeatherApp/app/src/main/java/dam_A51696/cooljetpackweatherapp/ui/WeatherUIState.kt
@@ -5,7 +5,7 @@ data class WeatherUIState(
     val longitude: Float = -9.1393f,
     val temperature: Float = 0f,
     val windspeed: Float = 0f,
-    val winddirection: Int = 0,
+    val winddirection: Float = 0f,
     val weathercode: Int = 0,
     val seaLevelPressure: Float = 0f,
     val time: String = ""
